requirements.txt
babel==2.17.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.0.3
flask-babel==4.0.0
gunicorn==22.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
